case

Trait CaseExt

pub trait CaseExt {
    type Owned;

    // Required methods
    fn to_capitalized(&self) -> Self::Owned;
    fn is_capitalized(&self) -> bool;
    fn to_camel(&self) -> Self::Owned;
    fn to_camel_lowercase(&self) -> Self::Owned;
    fn is_camel_lowercase(&self) -> bool;
    fn to_snake(&self) -> Self::Owned;
    fn to_dashed(&self) -> Self::Owned;
}

Required Associated Types

type Owned

Required Methods

fn to_capitalized(&self) -> Self::Owned

Create a new string from a string slice with replacing the first character with its to ASCII upper case counterpart (if one exists).

Examples

use case::CaseExt;

assert_eq!(&CaseExt::to_capitalized("stringy string"), "Stringy string");
assert_eq!(&CaseExt::to_capitalized("言語"), "言語");

fn is_capitalized(&self) -> bool

Check whether a string is capitalized.

If the the list is empty, the string is not considered capitalized.

Examples

use case::CaseExt;

assert_eq!("Stringy string".is_capitalized(), true);
assert_eq!("hello world".is_capitalized(), false);
assert_eq!("".is_capitalized(), false);

fn to_camel(&self) -> Self::Owned

Convert a string slice from snake case to a new capitalized camel case string.

Examples

use case::CaseExt;

assert_eq!(&"a_string_and_a_miss".to_camel(), "AStringAndAMiss");
assert_eq!(&"fooby".to_camel(), "Fooby");
assert_eq!(&"_wild__underscore_s_".to_camel(), "WildUnderscoreS");
assert_eq!(&"言_語".to_camel(), "言語");

fn to_camel_lowercase(&self) -> Self::Owned

Convert a string slice from snake case to a new uncapitalized camel case string.

Examples

use case::CaseExt;

assert_eq!(&"string_henry_iii".to_camel_lowercase(), "stringHenryIii");
assert_eq!(&"fooby".to_camel_lowercase(), "fooby");
assert_eq!(&"_wild__underscore_s_".to_camel_lowercase(), "wildUnderscoreS");
assert_eq!(&"言_語".to_camel_lowercase(), "言語");

fn is_camel_lowercase(&self) -> bool

Check whether a string is in camel lowercase.

If the the list is empty, the string is not considered camel lowercase.

Examples

use case::CaseExt;

assert_eq!("Stringy string".is_camel_lowercase(), false);
assert_eq!("helloWorld".is_camel_lowercase(), true);
assert_eq!("WildUnderscoreS".is_camel_lowercase(), false);
assert_eq!("".is_capitalized(), false);

fn to_snake(&self) -> Self::Owned

Convert a string from camel case to snake case.

Examples

use case::CaseExt;

assert_eq!(&"martinLutherStringJr".to_snake(), "martin_luther_string_jr");
assert_eq!(&"fooby".to_snake(), "fooby");
assert_eq!(&"WildUnderscoreS".to_snake(), "wild_underscore_s");
assert_eq!(&"言語".to_snake(), "言語");

fn to_dashed(&self) -> Self::Owned

Replaces underscores with dashes in a string.

Examples

use case::CaseExt;

assert_eq!(&"stringing_in_the_rain".to_dashed(), "stringing-in-the-rain");